Convert `@RenderPage()` to amendment
