+++ src/navigation/page-render/index.ts
@@ -1,4 +1,4 @@
 export * from './page-render-ctl';
 export * from './page-renderer';
-export * from './render-page.decorator';
+export * from './render-page.amendment';
 export * from './render-page-def';
